Rewrite get_model to use the device code
Get the S/N from /proc/usid, it's there on every device.
(Kindle 2 (possibly DX as well) detection was broken with the old code,
which probably means KOReader's detection is broken too).
Bonus: we get a finer-grained detection for the white Kindle DX ;).

See get_model.awk in resources for a commented version