Poor man's workaround for missing wraparound nav.
Bind [Aa] & [KEYBOARD] to a 'go to top' action, and [MENU] to a 'go to
bottom' action. It's mostly tailored for the K4 layout (where the
fiveway controller is framed by [KEYBOARD] on the left, and [MENU] on
the right, since we have the direct keybinds (1..0/Q..P) on the K2/K3.