Fix build w/ gawk 3
Switch from \s (which is a GNU extension only found in GNU Awk >= 4) to
the POSIX character class [:space:] in the strip awk script.
Adapt the version check accordingly, but keep it, because we still have
to rely on GNU Awk (mawk, bb awk & osx awk all fail to handle it in subtly
different ways).